agents/requirements.txt
yt-dlp
openai
python-dotenv
fastapi
uvicorn
requests
pydantic
